+++ src/main/java/insane96mcp/insanesurvivaloverhaul/mixin/module/combat/misccombat/PlayerMixin_MiscCombat.java
@@ -1,4 +1,4 @@
-package insane96mcp.insanesurvivaloverhaul.mixin.module.combat;
+package insane96mcp.insanesurvivaloverhaul.mixin.module.combat.misccombat;
 
 import com.llamalad7.mixinextras.expression.Definition;
 import com.llamalad7.mixinextras.expression.Expression;