+++ src/main/java/insane96mcp/insanesurvivaloverhaul/mixin/module/hungerhealth/healthregenhunger/PlayerMixin_HealthRegenHunger.java
@@ -1,4 +1,4 @@
-package insane96mcp.insanesurvivaloverhaul.mixin.module.hungerhealth;
+package insane96mcp.insanesurvivaloverhaul.mixin.module.hungerhealth.healthregenhunger;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import insane96mcp.insanesurvivaloverhaul.module.hungerhealth.HealthRegenHunger;