+++ src/main/java/insane96mcp/insanesurvivaloverhaul/module/mining/MiningMisc.java
@@ -3,12 +3,12 @@ package insane96mcp.insanesurvivaloverhaul.module.mining;
 import insane96mcp.insanelib.core.feature.Feature;
 import insane96mcp.insanelib.core.feature.LoadFeature;
 import insane96mcp.insanelib.core.feature.config.Config;
-import insane96mcp.insanesurvivaloverhaul.mixin.accessor.BlockStateBaseAccessor;
 import insane96mcp.insanesurvivaloverhaul.module.ISOModules;
-import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.item.DiggerItem;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.block.*;
+import net.minecraft.world.level.block.SlabBlock;
+import net.minecraft.world.level.block.StairBlock;
+import net.minecraft.world.level.block.WallBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.SlabType;
 import net.neoforged.bus.api.SubscribeEvent;
@@ -17,11 +17,6 @@ import net.neoforged.neoforge.event.entity.player.PlayerEvent;
 
 @LoadFeature(module = ISOModules.MINING, description = "Various mining changes")
 public class MiningMisc extends Feature {
-	@Config(name = "Insta-Mine Silverfish", description = "Silverfish blocks will insta-mine like pre-1.17")
-	public static Boolean instaMineSilverfish = true;
-	@Config(name = "Insta-Mine Heads", description = "Heads will insta-break")
-	public static Boolean instaMineHeads = true;
-
 	@Config(name = "Faster slabs, stairs and walls", description = "Makes slabs, stairs and walls take less time to break")
 	public static Boolean fastSlabsStairsWalls = true;
 	@Config(description = "In vanilla there's a 5 tick delay (0.25 secs) between breaking blocks. The tick delay is reduced by 1 tick every 2.5 tool mining speed.")
@@ -32,11 +27,6 @@ public class MiningMisc extends Feature {
 	@Override
 	public void readConfig(ModConfigEvent event) {
 		super.readConfig(event);
-		for (Block block : BuiltInRegistries.BLOCK.stream().toList()) {
-			if ((instaMineHeads && block instanceof AbstractSkullBlock)
-					|| (instaMineSilverfish && block instanceof InfestedBlock))
-				block.getStateDefinition().getPossibleStates().forEach(blockState -> ((BlockStateBaseAccessor) blockState).setDestroySpeed(0f));
-		}
 		/*if (ModList.get().isLoaded("caverns_and_chasms") && cavernsChasmsIntegration) {
 			CCConfig.COMMON.chainmailArmorIncreasesDamage.set(false);
 			CCConfig.COMMON.goldenArmorIncreasesSpeed.set(false);